+++ src/interfaces.ts
@@ -3,3 +3,10 @@ export const enum NumeralSystem {
 	Hexadecimal = 2,
 	Roman = 3
 }
+
+export const enum NumberArithmetic {
+	SixtyFourBit = 1,
+	ThirtyTwoBit = 2,
+	SixteenBit = 3,
+	EightBit = 4
+}